tasks/prepare-opa-properties/requirements.txt
functions-framework==3.*
pandas==2.*
pyarrow==15.*
google-cloud-storage==2.*
